[ltsmaster] std.math.pow: change type from `double` to `real` for internal variable.
This change fixes a unittest failure with LDC with optimizations on. Note that the return type of `impl` is `real` (not `Float`) and thus changing the type of `sign` to `real` should not matter / be more correct? I am not sure why `w` on line 6299 is of type `Float`.

(cherry picked from commit 4b1f2f0147cfa8a6fed9bc9830f72060024d305d)